Update patch set 1
Patch Set 1:

I'm happy to cherrypick this to internal master, if we still think that's the right direction. That's where it would have to be merged first. I think it would have to be manually cherrypciked to 2 or 3 other branches as well.

Making this header-only, as enh suggests, would probably make this easier to merge, though it seems a little strange.

This had been on my list, but had gotten deprioritized, since my hope of an easier merge after pi release evaporated.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>